refactor: move and rename dropdown atoms
